moving things into components
