+++ src/renderer/components/content-types/workspace/TitleBar.tsx
@@ -2,7 +2,7 @@ import React, { useRef, useState, useEffect } from 'react'
 import { clipboard } from 'electron'
 
 import Dropdown, { DropdownContent, DropdownTrigger } from 'react-simple-dropdown/dropdown'
-import Omnibox from './Omnibox'
+import Omnibox from './omnibox/Omnibox'
 import Content from '../../Content'
 import Authors from './Authors'
 import { HypermergeUrl, PushpinUrl } from '../../../ShareLink'
