+++ src/renderer/components/content-types/workspace/omnibox/Action.css
