+++ src/renderer/components/content-types/workspace/omnibox/Action.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import SecondaryText from '../../SecondaryText'
+import SecondaryText from '../../../SecondaryText'
 import './Action.css'
 
 interface Props {
