+++ src/renderer/components/content-types/workspace/omnibox/Omnibox.css
